# api/openapi.yaml
openapi: 3.0.0
info:
  title: Matrix Computation API
  description: API to compute a matrix of double values from input parameters.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local server
  - url: http://azakan.net:5000
    description: Remote server
paths:
  /AnalyticalSigmaVolatilityCalibration:
    post:
      summary: Compute a matrix from input parameters
      description: Takes various input parameters and returns computed values.
      operationId: volatility_densityPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyticalSigmaVolatilityCalibration_Input'
        required: true
      responses:
        "200":
          description: Successful computation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticalSigmaVolatilityCalibration_Output'
        "400":
          description: Invalid input parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "500":
          description: Internal server error or computation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-swagger-router-controller: Default
...
  /api/Lognormal_FX_With_MHJM_Rates:
    get:
      summary: Get HJM calibration results with lognormal FX rates
      operationId: LognormalFXWithMHJMRates
      x-swagger-router-controller: Default
      parameters:
        - name: num_paths
          in: query
          required: false
          schema:
            type: integer
            default: 524288
            description: Number of simulation paths
        - name: volatility
          in: query
          required: false
          schema:
            type: number
            default: 0.3
            description: Initial volatility parameter
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [success, error]
                  data:
                    type: object
                    properties:
                      strikes:
                        type: array
                        items:
                          type: number
                      model_volatility:
                        type: array
                        items:
                          type: number
                      market_volatility:
                        type: array
                        items:
                          type: number
                      results_money_market:
                        type: array
                        items:
                          type: number
                      results_discount_factors:
                        type: array
                        items:
                          type: number
                  error:
                    type: string
                    nullable: true
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/analytical-sigma-volatility:
    get:
      summary: Get analytical sigma volatility calculations
      operationId: getAnalyticalSigmaVolatility
      x-swagger-router-controller: Default
      parameters:
        - name: n
          in: query
          schema:
            type: integer
            default: 200
        - name: fwd
          in: query
          schema:
            type: number
            default: 2245.0656707892695  # Default for Test = 1
        - name: time
          in: query
          schema:
            type: number
            default: 1.0  # Default for Test = 1
        - name: ctrl_p
          in: query
          schema:
            type: number
            default: 0.2
        - name: ctrl_c
          in: query
          schema:
            type: number
            default: 0.2
        - name: atm
          in: query
          schema:
            type: number
            default: 1.1  # Default for Test = 1
        - name: skew
          in: query
          schema:
            type: number
            default: 3.5  # Default for Test = 1
        - name: smile
          in: query
          schema:
            type: number
            default: 17  # Default for Test = 1
        - name: put
          in: query
          schema:
            type: number
            default: 0.7  # Default for Test = 1
        - name: call
          in: query
          schema:
            type: number
            default: 0.06  # Default for Test = 1
        - name: Test
          in: query
          schema:
            type: integer
            default: 1
            description: |
              Test case selector:
              1: High values case
              2-4: Standard values (fwd=1.0, time=0.333, atm=0.1929, skew=0.02268, 
                 smile=0.00317, put=-0.00213, call=-0.00006)
      responses:
        '200':
          $ref: '#/components/schemas/AnalyticalSigmaVolatilityResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
    AnalyticalSigmaVolatilityCalibration_Input:
      type: object
      required:
        - n
        - spot
        - expiry
        - r
        - q
        - beta
        - rho
        - volvol
        - computationType
      properties:
        n:
          type: integer
          description: Number of points
          minimum: 1
          example: 400
        spot:
          type: number
          format: double
          description: Spot price
          minimum: 0
          exclusiveMinimum: true
          example: 2273.684211
        expiry:
          type: number
          format: double
          description: Time to expiry in years
          minimum: 0
          exclusiveMinimum: true
          example: 1.0
        r:
          type: number
          format: double
          description: Risk-free interest rate
          example: 0.003
        q:
          type: number
          format: double
          description: Dividend yield
          minimum: 0
          example: 0.0022
        beta:
          type: number
          format: double
          description: Beta parameter
          minimum: 0
          maximum: 1
          example: 0.4158
        rho:
          type: number
          format: double
          description: Correlation parameter
          minimum: -1
          maximum: 1
          example: 0.2256
        volvol:
          type: number
          format: double
          description: Volatility of volatility
          minimum: 0
          exclusiveMinimum: true
          example: 0.2256
        computationType:
          type: string
          description: Type of computation to perform
          enum:
            - volatility_asv
            - density
            - volatility_svi
          example: volatility_asv
    AnalyticalSigmaVolatilityCalibration_Output:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
          description: Operation status
        data:
          type: array
          items:
            type: object
            properties:
              key:
                type: number
                format: double
                description: Input value (e.g., strike price)
              value:
                type: number
                format: double
                description: Computed result
            required:
              - key
              - value
        error:
          type: string
          nullable: true
          description: Error message if status is error
      required:
        - status
    AnalyticalSigmaVolatilityResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, error]
          description: Operation status
        data:
          type: object
          properties:
            strikes:
              type: array
              items:
                type: number
              description: Array of strike prices
            vols:
              type: array
              items:
                type: number
              description: Array of volatilities
            vols_origin:
              type: array
              items:
                type: number
              description: Array of origin volatilities
        error:
          type: string
          nullable: true
          description: Error message if status is error
...
    Error:
      type: object
      required:
        - status
        - error
      properties:
        status:
          type: string
          enum: [error]
          description: Error status indicator
        error:
          type: string
          description: Detailed error message
        timestamp:
          type: string
          format: date-time
          description: Timestamp of when the error occurred
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'